03_agents/00_agency_swarm_framework/03_requirement_generator_agent/requirements.txt
python-docx 
Pillow 
markdown